Step refactor

+++ src/main/java/vn/id/milease/mileaseapi/service/impl/PlanServiceImpl.java
@@ -10,9 +10,11 @@ import vn.id.milease.mileaseapi.model.dto.create.CreatePlanDto;
 import vn.id.milease.mileaseapi.model.dto.search.PlanSearchDto;
 import vn.id.milease.mileaseapi.model.dto.update.UpdatePlanDto;
 import vn.id.milease.mileaseapi.model.entity.plan.Plan;
+import vn.id.milease.mileaseapi.model.entity.user.User;
 import vn.id.milease.mileaseapi.model.exception.NotFoundException;
 import vn.id.milease.mileaseapi.repository.PlanRepository;
 import vn.id.milease.mileaseapi.service.PlanService;
+import vn.id.milease.mileaseapi.service.UserService;
 import vn.id.milease.mileaseapi.service.util.ServiceUtil;
 import vn.id.milease.mileaseapi.util.ApplicationMapper;
 
@@ -23,17 +25,19 @@ import javax.transaction.Transactional;
 @Transactional
 public class PlanServiceImpl implements PlanService {
     private final PlanRepository planRepository;
+    private final UserService userService;
     private final ApplicationMapper mapper;
 
     @Override
     public PlanDto getPlanById(long id) {
-        // TODO [Duy, P1] check user permission
-        return mapper.getPlanMapper().toDto(getPlan(id));
+        Plan plan = getPlan(id);
+        checkCurrentUserPermission(plan);
+        return mapper.getPlanMapper().toDto(plan);
     }
 
     @Override
     public PageResult<PlanDto> getPlans(PlanSearchDto searchDto) {
-        // TODO [Duy, P1] check user permission
+        // TODO [Duy, P1] only allow user to get their own plans
         var predicate = planRepository.prepareSearchPredicate(searchDto);
         PageRequest pageRequest = ServiceUtil.preparePageRequest(searchDto);
         Page<Plan> plans = planRepository.findAll(predicate, pageRequest);
@@ -42,7 +46,7 @@ public class PlanServiceImpl implements PlanService {
 
     @Override
     public PlanDto addPlan(CreatePlanDto dto) {
-        // TODO [Duy, P1] check user permission
+        // TODO [Duy, P1] link user when create new plan
         Plan entity = mapper.getPlanMapper().toEntity(dto);
         // TODO [Duy, P2] validate that the start must be after today and end date must be after start date
         entity.setId(0L);
@@ -52,8 +56,8 @@ public class PlanServiceImpl implements PlanService {
 
     @Override
     public PlanDto updatePlan(UpdatePlanDto dto) {
-        // TODO [Duy, P1] check user permission
         Plan entity = getPlan(dto.getId());
+        checkCurrentUserPermission(entity);
         // TODO [Duy, P2] validate status
         mapper.getPlanMapper().toEntity(dto, entity);
         entity = planRepository.save(entity);
@@ -62,10 +66,16 @@ public class PlanServiceImpl implements PlanService {
 
     @Override
     public void deletePlan(long id) {
-        // TODO [Duy, P1] check user permission
-        var entity = getPlan(id);
+        Plan entity = getPlan(id);
+        checkCurrentUserPermission(entity);
         // TODO [Duy, P1] improve this
         planRepository.delete(entity);
+    }
+
+    @Override
+    public void checkCurrentUserPermission(Plan plan) {
+        User user = userService.getCurrentUser();
+        // TODO [Duy, P1] Check user permission
     }
 
     public Plan getPlan(long id) {
